lo del componente mainAdvertising se movio al componente index del aviso
