--- src/components/AdminWeb/Pages/Avisos/components/ButtonCreateAdvertising.tsx
@@ -1,18 +1,0 @@
-type ButtonCreateadvertisingProps = {
-  onClick: () => void;
-};
-
-function ButtonCreateAdvertising({ onClick }: ButtonCreateadvertisingProps) {
-  return (
-    <>
-      <button
-        onClick={onClick}
-        className="bg-[#2C9CBF] rounded-[15px] select-none py-[16px] w-[236px] text-white font-[600] text-[20px]"
-      >
-        NUEVO AVISO
-      </button>
-    </>
-  );
-}
-
-export default ButtonCreateAdvertising;
